refactor: move tests related to data access
